nonbug-bugfix: insertion of new headline title
In the original code, if there is whitespace on the last line of the file, then
the headline *should* insert after the whitespace, but it doesn't!  I don't
understand the magic that makes this happen.

In this bugfix, we look for `\n` as the last character of the file and add `\n`
before "* <title>" if the last character is anything else.  This requires no
magic to work.